User datasets tweaks (#1193)
* Verbiage updates, and make description optional by default

* Remove link to user dataset workspace from study summaries page

* Make spacing between elements more consistent

* Reorder attributes on detail page, remove quota usage, and refactor some page sections

* Refactor

(cherry picked from commit eb7af83dbfd152ccc3a6997ab811894c8233adcd)
