+++ packages/libs/user-datasets/src/lib/Components/Detail/BigwigDatasetDetail.jsx
@@ -1,7 +1,11 @@
 import React from 'react';
 
 import Icon from '@veupathdb/wdk-client/lib/Components/Icon/IconAlt';
-import { Mesa, MesaState } from '@veupathdb/coreui/lib/components/Mesa';
+import {
+  AnchoredTooltip,
+  Mesa,
+  MesaState,
+} from '@veupathdb/coreui/lib/components/Mesa';
 
 import { makeClassifier } from '../UserDatasetUtils';
 import UserDatasetDetail from './UserDatasetDetail';
@@ -14,6 +18,11 @@ class BigwigDatasetDetail extends UserDatasetDetail {
     super(props);
     this.renderTracksSection = this.renderTracksSection.bind(this);
     this.getTracksTableColumns = this.getTracksTableColumns.bind(this);
+    this.renderCompatibilitySection =
+      this.renderCompatibilitySection.bind(this);
+    this.getCompatibilityStatus = this.getCompatibilityStatus.bind(this);
+    this.getCompatibilityTableColumns =
+      this.getCompatibilityTableColumns.bind(this);
     this.state = {
       ...this.state,
       sequenceId: null,
@@ -133,11 +142,135 @@ class BigwigDatasetDetail extends UserDatasetDetail {
     );
   }
 
+  /* -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
+
+                                Compatible Table
+
+   -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-= */
+
+  renderCompatibilitySection() {
+    const { userDataset, config, dataNoun } = this.props;
+    const { displayName } = config;
+
+    const compatibilityTableState = MesaState.create({
+      columns: this.getCompatibilityTableColumns(userDataset),
+      rows: userDataset.dependencies,
+    });
+
+    const compatibilityStatus = this.getCompatibilityStatus();
+
+    return (
+      <section id="dataset-compatibility">
+        <h2>
+          Use This {dataNoun.singular} in {displayName}
+        </h2>
+        <h3>
+          <Icon fa="puzzle-piece" />
+          Compatibility Information &nbsp;
+          <AnchoredTooltip
+            content={`The data and genomes listed here are requisite for using the data in this user ${dataNoun.singular.toLowerCase()}.`}
+          >
+            <div className="HelpTrigger">
+              <Icon fa="question-circle" />
+            </div>
+          </AnchoredTooltip>
+        </h3>
+        <div style={{ maxWidth: '600px' }}>
+          <Mesa state={compatibilityTableState} />
+        </div>
+        {compatibilityStatus}
+      </section>
+    );
+  }
+
+  getCompatibilityStatus() {
+    const { userDataset, config, dataNoun } = this.props;
+    const { projectId } = config;
+
+    const { status, projects } = userDataset;
+
+    /**
+     * In VDI, we know a dataset is compatible when the site-specific's install status
+     * indicates a successful install.
+     *
+     * We know a dataset is incompatible when the site-specific's install status
+     * indicates `missing-dependency`
+     */
+    const installStatusForCurrentProject = status.install?.find(
+      (d) => d.projectId === projectId
+    );
+
+    const isTargetingCurrentSite = projects.includes(projectId);
+    const isInstalled = [
+      userDataset.status.import,
+      installStatusForCurrentProject?.metaStatus,
+      installStatusForCurrentProject?.dataStatus,
+    ].every((status) => status === 'complete');
+
+    const isIncompatible =
+      installStatusForCurrentProject?.dataStatus === 'missing-dependency';
+
+    if (!isTargetingCurrentSite || (isTargetingCurrentSite && isIncompatible)) {
+      return (
+        // if projectIds don't match, then we're not installable and thus incompatible
+        // if we're installable but failed due to a missing dependency, we're incompatible
+        <p className="danger">
+          This {dataNoun.singular.toLowerCase()} is not compatible with{' '}
+          <b>{projectId}</b>.
+        </p>
+      );
+    } else if (isInstalled) {
+      return (
+        // if we've installed successfully and we're installable, we're compatible
+        <p className="success">
+          This {dataNoun.singular.toLowerCase()} is compatible with{' '}
+          <b>{projectId}</b>. It is installed for use.
+        </p>
+      );
+    } else {
+      // instead of attempting to provide very granular messaging for when things are neither
+      // compatible nor incompatible, let's let the dataset page's Status messaging handle this
+      return null;
+    }
+  }
+
+  getCompatibilityTableColumns() {
+    const { userDataset } = this.props;
+    const { projects } = userDataset;
+    return [
+      {
+        key: 'project',
+        name: 'VEuPathDB Website',
+        renderCell() {
+          return projects.join(', ');
+        },
+      },
+      {
+        key: 'resourceDisplayName',
+        name: 'Required Resource',
+        renderCell({ row }) {
+          const { resourceDisplayName } = row;
+          return resourceDisplayName;
+        },
+      },
+      {
+        key: 'resourceVersion',
+        name: 'Required Resource Release',
+        renderCell({ row }) {
+          const { resourceVersion } = row;
+          return resourceVersion;
+        },
+      },
+    ];
+  }
+
+  // See note in the base class, UserDatasetDetail
+  /** @return {import("react").ReactNode[]} */
   getPageSections() {
-    const [headerSection, compatSection, fileSection] = super.getPageSections();
+    const [headerSection, fileSection] = super.getPageSections();
     return [
       headerSection,
-      compatSection,
+      this.renderCompatibilitySection,
       this.renderTracksSection,
       fileSection,
     ];
