+++ packages/libs/user-datasets/src/lib/Components/Detail/UserDatasetDetail.scss
@@ -24,8 +24,16 @@
     padding: 20px 0;
   }
 
-  .UserDatasetDetail-SectionTitle {
-    margin-bottom: 10px;
+  h2 {
+    padding: 0;
+    font-size: 1.65em;
+    color: black;
+  }
+  h3 {
+    padding: 0;
+    font-size: 1.3em;
+    margin-top: 1.1em;
+    margin-bottom: 0.5em;
     .wdk-Icon {
       color: $blue;
       margin-right: 10px;
@@ -81,6 +89,14 @@
       font-size: 1.4em;
       font-weight: 300;
 
+      &:last-child {
+        margin-bottom: 0;
+      }
+
+      a {
+        font-weight: 500;
+      }
+
       &.UserDatasetDetail-Name {
         font-weight: 400;
         font-size: 2.8em;
@@ -100,18 +116,23 @@
       }
       .UserDatasetDetail-AttributeValue {
         flex: 1 1 auto;
+        ul {
+          display: flex;
+          list-style: none;
+          margin: 0;
+          li + li:before {
+            content: '|';
+            margin: 0 1ex;
+          }
+        }
       }
     }
-    .UserDatasetDetail-Owner {
-      margin-top: 1em;
-    }
-    .UserDatasetDetail-Summary {
-      margin-top: 2em;
-      margin-bottom: 1em;
-      flex-direction: column;
-    }
-    .UserDatasetDetail-Description {
-      flex-direction: column;
+    .UserDatasetDetail {
+      &-Owner,
+      &-Summary,
+      &-Created {
+        margin-top: 1em;
+      }
     }
   }
 }
