+++ packages/libs/user-datasets/src/lib/Components/UserDatasets.scss
@@ -23,7 +23,7 @@
 
   section {
     width: 100%;
-    margin-bottom: 30px;
+    margin-bottom: 1.75em;
   }
 
   .MesaComponent {
